reorganizacion de el package screens
