+++ app/src/androidTest/java/ar/edu/unlam/mobile/scaffolding/ui/components/EventSearchBarTest.kt
@@ -5,8 +5,8 @@ import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performTextInput
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import ar.edu.unlam.mobile.scaffolding.ui.common.EventSearchState
-import ar.edu.unlam.mobile.scaffolding.ui.screens.SearchUIState
+import ar.edu.unlam.mobile.scaffolding.ui.screens.home.SearchUIState
+import ar.edu.unlam.mobile.scaffolding.ui.screens.home.state.EventSearchState
 import org.junit.Assert
 import org.junit.Rule
 import org.junit.Test
