+++ app/src/main/java/ar/edu/unlam/mobile/scaffolding/ui/theme/Color.kt
@@ -8,3 +8,5 @@ val LightOlive = Color(0xFFF4FCE6)
 val Green = Color(0xFF658C58)
 val LightGreen = Color(0xFFFCFFF7)
 val DarkGreen = Color(0xFF31694E)
+val MintGreen = Color(0xFF9EEBC2)
+val DeepGreen = Color(0xFF1B5039)
