+++ app/src/main/java/ar/edu/unlam/mobile/scaffolding/ui/theme/Theme.kt
@@ -13,26 +13,24 @@ import androidx.compose.ui.platform.LocalContext
 
 val DarkColorScheme =
     darkColorScheme(
-        primary = Color(0xFF8FCF84),
-        onPrimary = Color(0xFF0D3709),
-        primaryContainer = Color(0xFF456B3B),
-        onPrimaryContainer = Color(0xFFC9F0BE),
+        primary = Green,
+        onPrimary = Color.White,
+        primaryContainer = Color(0xFF2B2F29),
+        onPrimaryContainer = Color.White,
         secondary = Color(0xFFD7DB8E),
         onSecondary = Color(0xFF34360A),
         secondaryContainer = Color(0xFF4D4F24),
         onSecondaryContainer = Color(0xFFEFEFB6),
-        tertiary = Color(0xFF5BB28A),
-        onTertiary = Color(0xFF003823),
+        tertiary = DarkGreen,
+        onTertiary = Color.White,
         tertiaryContainer = Color(0xFF1B5039),
         onTertiaryContainer = Color(0xFF9EEBC2),
         error = Color(0xFFCF6679),
         onError = Color.Black,
         errorContainer = Color(0xFF8C1D18),
         onErrorContainer = Color(0xFFF9DEDC),
-        background = Color(0xFF121312),
+        background = Color(0xFF1D1F1D),
         onBackground = Color(0xFFE1E3E0),
-        surface = Color(0xFF1B1C1A),
-        onSurface = Color(0xFFE1E3E0),
         surfaceVariant = Color(0xFF2D322B),
         onSurfaceVariant = Color(0xFFC3C8BE),
         outline = Color(0xFF484E45),
@@ -55,10 +53,12 @@ private val LightColorScheme =
         onSecondaryContainer = Color.Black,
         tertiary = DarkGreen,
         onTertiary = Color.White,
+        tertiaryContainer = DeepGreen,
+        onTertiaryContainer = MintGreen,
         background = Color.White,
         onBackground = DarkGreen,
-        surface = LightOlive,
         onSurface = DarkGreen,
+        error = Color(0xFFCF6679),
     )
 
 @Composable
